+++ src/action/index.js
@@ -1,13 +1,14 @@
 import axios from 'axios';
 
-//bring in user input for difficulty and category
+export const quizSettings = (...settings) => ({ type: 'UPDATE_SETTINGS', payload: settings })
+
+export const localScores = (scores) => ({ type: 'UPDATE_LOCAL_SCORES', payload: scores })
+
 export const fetchCategories = () => {
     return async (dispatch) => {
         try {
-            const { data } = await axios.get('https://opentdb.com/api.php?amount=10&category=22&difficulty=easy&type=multiple')
-            console.log(data.results);
-            let newCategoryArray = data.results.map((data, i) => ({ id: i + 1, category: data.category}))
-            console.log(newCategoryArray);
+            const { data } = await axios.get('https://opentdb.com/api_category.php')
+            let newCategoryArray = data.trivia_categories.map((data) => ({id: data.id, category: data.name}))
             dispatch ({
                 type: 'UPDATE_CATEGORY',
                 payload: newCategoryArray
@@ -17,6 +18,29 @@ export const fetchCategories = () => {
                 type: 'SET_ERROR',
                 payload: err
             })
+        }
+    }
+}
+
+export const fetchQuestions = (category, level) => {
+    return async (dispatch) => {
+        dispatch({type:'LOADING', payload: true})
+        try {
+            const { data } = await axios.get(`https://opentdb.com/api.php?amount=10&category=${category}&difficulty=${level}&type=multiple`)
+            if(data.results.length==0){ throw Error('Not enough questions in that category please pick a different one')}
+            let newQuestionArray = data.results
+            dispatch ({
+                type: 'UPDATE_QUESTIONS',
+                payload: newQuestionArray,
+            })
+        } catch (err) {
+            
+            console.warn(err.message)
+            dispatch ({
+                type: 'SET_ERROR',
+                payload: err.message,
+            })
+            throw Error(err.message);
         }
     }
 }