+++ src/components/index.js
@@ -1,2 +1,4 @@
 export { default as AddUser } from './AddUser';
-export { default as useFetch } from './useFetch';
+export { default as Question } from './Question';
+export { default as Timebar } from './Timebar'
+export { default as Stars } from './Stars'