+++ src/layout/Modal/index.js
@@ -1,15 +1,16 @@
-import React, { useState } from 'react';
-import AddUser from '../../components/AddUser/addUser'
+import React, { useState, useEffect } from 'react';
+import {AddUser} from '../../components/'
 import './style.css'
 
 function Modal( { getResults, show }) {
+    const [users, setUsers] = useState([])
+
+    useEffect(()=>{
+        console.log(users);
+    },[])
 
     const handler = (e) => {
         e.preventDefault();
-        let users=[]
-        for(let x of e.target){
-            if(x.value) users.push(x.value);
-        }
         getResults(users)
         show();
     }
@@ -17,11 +18,11 @@ function Modal( { getResults, show }) {
     return (
         <div className='userModal'>
             <div>
-                <span onClick={show}>x</span>
-                <form onSubmit={handler}>       
-                    <AddUser toggle={false} first={false}/>
-                    <input type="submit"/>
-                </form>
+                <span onClick={show}  style={{cursor: "pointer"}}>x</span>
+                <div className='modalContent'>       
+                    <AddUser toggle={false} first={false} returnedValue={(e)=>setUsers(e)}/>
+                    <button onClick={handler}>submit</button>
+                </div>
             </div>
         </div>
     )