+++ src/layout/index.js
@@ -1,1 +1,3 @@
-export { Modal } from './Modal'
+export { default as Modal } from './Modal'
+export { default as Footer } from './Footer'
+export { default as Header } from './Header'